reorganize project to be more general
