+++ wireguard/keys.go
@@ -19,7 +19,8 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 */
-package keys
+
+package wireguard
 
 import (
 	"golang.zx2c4.com/wireguard/wgctrl/wgtypes"
